requirements.txt
decorator==5.1.0
greenlet==1.1.2
numpy==1.21.4
pandas==1.3.4
pbr==5.8.0
psycopg2==2.9.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.27
sqlalchemy-migrate==0.13.0
sqlparse==0.4.2
Tempita==0.5.2
